+++ ninja-core/src/test/java/ninja/utils/SimpleClass.java
@@ -1,0 +1,11 @@
+package ninja.utils;
+
+/**
+ * Simple class. Used to test e.g. the {@link ObjectMapper}.
+ */
+public class SimpleClass {
+    
+    public String name;
+    public String test;
+    
+}