+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,12 +1,15 @@
-Version 1.0.X
+Version 1.1
 =============
  
+ * Bugfix: explicitly set content type of result keeps untouched by ResultHandler now (roughy/hennig)
+ * Added HttpOnly option for the session cookie (roughy)
+ * Added support for simple parsing of post forms (aka application/x-www-form-urlencoded) (zoza)
  * PLZ ADD YOUR CHANGES HERE :)
  
 Version 1.0.8
 =============
  
- * Fixed bug + test with flash scope: https://github.com/reyez/ninja/issues/70 (ra)
+ * Fixed bug + test with flash scope: https://github.com/reyez/ninja/issues/70 (zoza / ra)
 
 Version 1.0.7
 =============